secp256k1-bitcoinj: Move WitnessMaker to src/main from src/test
