+++ secp256k1-bitcoinj/src/main/java/org/bitcoinj/secp256k1/bitcoinj/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This module contains experimental utility classes for integrating {@code secp256k1-jdk} with <b>bitcoinj</b>.
+ * If/when a future version of bitcoinj supports secp256k1-jdk and/or Schnorr signatures, this package will likely
+ * no longer be needed and will be deprecated.
+ */
+package org.bitcoinj.secp256k1.bitcoinj;
